==> R Code/02_NSFG_cleaning_F.Rmd <==
---
title: "NSFG_Cleaning"
author: "Jennifer Spencer"
date: "9/22/2019"
output: html_document
---


This file imports 4 cycles (8 years worth) of data from the National Survey of Family Growth

Data: 2011-2017

**The file proceeds as follows:**

**[1] We import the relevant files from raw NSFG files.** 

**[2] We create a matched dataset containing all years and only the variables of interest.**

**[3] We define sexual behavior categories based on survey data **

**[4] Define several target statistics and validation measure for our model. ** 

```{r setup, include=FALSE}
# First we clear the system and load packages
rm(list=ls())

library(tidyverse)   #ggplot & data manipulation packages
library(survey)      #for incorporating complex survey design
library(kableExtra)  #for making tables look nice
library(ggpubr)      #for making figures look nice
library(readstata13) #for reading in the Stata Files
library(openxlsx)    #For writing to excel files

```

## Import Files

The NSFG provides Stata or SAS read-in files along with data dictionaries. A separate piece of Stata code (NSFG_Inputs.do) reads in all the raw files and saves as .dta. 

Here I read in these files as .dta and save only the relevant variables 

####Female Data
    Detailed questions asked about sex with men, but fewer about sex with women
    **Demographics** Age and ID
    **Age of Sexual Onset**  Questions about first age at first sex
    **M Partner Data** Have you ever had sex with a man, how many ever, how many past year
    **Egocentric Heterosexual Data** Detailed information about first relationship and up to three past-year relationships
    **CASI Data M Partner** Additional questions on whether any M partners have also had sex with men and whether any have been in concurrent relationships
    **F Partner data** Have you ever had sex with a woman, how many ever, how many past year


```{r F_function, warning=FALSE}

f.var_F_data       <- function(data, year) {
data_out           <- data   %>% select(          
                      # # # # Demographics   # # # # 
                                                   CASEID,    #ID
                                            "strata"=SEST,    #Strata Variable
                                             "panel"=SECU,    #Panel Variable
                                               "age"=AGER,    #Age
                                               
                      # # # #  Married and Cohab w/ Man   # # # #  
                                      "marital_stat"= FMARITAL,    #Marital Status
                                        "cohab_stat"=COHEVER,    #Cohabitation Status
                      
                      # # # # Age of Onset   # # # #   
                                        "age_first"=AGEFSTSX,    #what age were you when you first had sex
                                        "age_same"=SAMESEX1,    #age at first same-sex experience

                      # # # #  Numbers of M Partners   # # # #   
                                          "eversex"=RHADSEX,     #have you ever had sex ---with a man---
                                       "num_m_ever"=LIFEPRT,     #total number of partners 
                                   "num_m_range_lo"=LIFEPRT_LO,  #If range of numbers was given
                                   "num_m_range_hi"=LIFEPRT_HI,  #If range of numbers was given
                                         "num_m_yr"=MON12PRT,    #number of partners in past year
                                      "num_m_yr_lo"=MON12PRT_LO, #If range of ages was given
                                      "num_m_yr_hi"=MON12PRT_HI, #If range of ages was given
                                           "num_yr"=PARTS1YR,    #Number of partners in past 12 months
                                         "num_curr"=CURRPRTT,    #number of current male partners
                      
                                        
                      # # # #  First Male Partner    # # # #                 
                             #          "p_age_rel"=FPRELAGE ,   #First partners age relative to womans (older/younger)
                             #         "p_age_diff"=FPRELYRS,    #number of years of age difference (older/younger)
                                  "current_p_first"=STILFPSX,    #In current relationship with first partner
                                     "last_p_first"=CMLSEXFP,    #date last had sex with first ever partner
                                       "first_same"=SAMEMAN,     #first partner already discussed 
                                        "who_same"=WHOFSTPR,     #identity of first partner 
                      
                      # # # #  Most recent Male(?) Partner    # # # #                                                                  
                                         "pr_check"=MATCHFP,     #is most recent partner first partner? 
                                         "hp_check"=MATCHHP,     #most recent partner is an already discussed partner
                                    "last_p_recent"=LSEXDATE,    #date of last sex with most recent partner
                                   "first_p_recent"=CMFSEX,      #date of first sex with most recent partner
                                 "current_p_recent"=PCURRNT,    #Is this partner a current partner? 
                                   "ageat_p_recent"=P1YRAGE,     #Age at first sex with most recent partner       
                                    "page_p_recent"=P1YHSAGE,    #partner's age at first sex with most recent partner 
                                     "rel_p_recent"=P1YRELP,     #relationship to most recent sexual partner
                                    
                            # # # #  2nd most recent Male Partner    # # # #                              
                                         "p2_check"=MATCHFP2,    #is second-most recent partner first partner? 
                                          "last_p2"=CMLSEX2,     #date of most last sex with 2nd most recent partner
                                         "first_p2"=CMFSEX2,     #date of first sex with 2nd most recent partner
                                       "current_p2"=PCURRNT2,    #is 2nd most recent partner a current partner?  
                                         "ageat_p2"=P1YRAGE2,    #Age at first sex with 2nd most recent partner       
                                          "page_p2"=P1YHSAGE2,   #partner's age at first sex with 2nd most recent partner 

                          # # # #  3rd most recent Male Partner    # # # #   
                                         "p3_check"=MATCHFP3,    #is third most recent partner first partner? 
                                          "last_p3"=CMLSEX3,     #date of most last sex with 3rd most recent partner
                                         "first_p3"=CMFSEX2,     #date of first sex with 3rd most recent partner
                                       "current_p3"=PCURRNT3,    #is 3rd most recent partner a current partner? 
                                         "ageat_p3"=P1YRAGE3,    #Age at first sex with 3rd most recent partner       
                                          "page_p3"=P1YHSAGE3,   #partner's age at first sex with 3rd most recent partner 
                             
 
                          # # # #  CASI Data -M Partners    # # # #                                             
                                #     "num_m_ever2"=PARTSLIF_1,   #Male Partners ever- asked again  #Same as above so removed
                                #       "num_m_any"=OPPLIFENUM,   #Male partners for any kind of sexual contact (not just vaginal)
                                #       "num_m_yr2"=PARTS12M_1,   #Male Partners past year- asked again
                                   "yr_bi_partner"=BISEXPRT,      #Any MSM partners
                                 "yr_nonm_partner"=NONMONOG,      #Any non-monogomous partners (male?)
                                "num_nonm_partner"=NNONMONOG1,     #Number non-monogomous partners  (male?)   
                                  
                          # # # #  CASI Data -F Partners    # # # #                      
                                          "ever_f"=SAMESEXANY,   #Female Partners ever
                                      "num_f_ever"=SAMLIFENUM,   #Number F partners lifetime
                                        "num_f_yr"=SAMYEARNUM,   #Number F partners in past year
                                   "last_p_gender"=MFLASTP,      #Gender of last partner M/F (if multiple sex partners in last year)
                                      "duration_1"=PARTDUR1,     #duration of most recent partnership
                                      "duration_2"=PARTDUR2,     #duration of second to last partnership
                                      "duration_3"=PARTDUR3,     #duration of third to last partnership
                                   )

   if(year==2017) { 
data_orientation      <- data   %>% select( 
                                                  CASEID,
                                    "orientation"=ORIENT_A        #self-described orientation    
                                    )

} else { 
data_orientation      <- data   %>% select( 
                                                  CASEID,
                                    "orientation"=ORIENT         #self-described orientation    
                                    )

}

data_out             <- merge(data_out, data_orientation, by="CASEID", all = TRUE)


data_out$year   <- year
data_out$gender <- "F"

return(data_out)
}


```

Then use this loop to pull in all these variables 

```{r all_years_F, warning=FALSE}

#Read in raw data
F_2013             <- read.dta13("./_2a_Inputs/NSFG/2011-2013/2011_2013.FemResp.dta")
F_2015             <- read.dta13("./_2a_Inputs/NSFG/2013-2015/2013-2015_FemResp.dta")
F_2017             <- read.dta13("./_2a_Inputs/NSFG/2015-2017/2015_2017_FemResp.dta")

#clean 2013
F_2013_vars        <- f.var_F_data(F_2013, 2013)

#clean 2015
F_2015_vars        <- f.var_F_data(F_2015, 2015)

#clean 2017
F_2017_vars        <- f.var_F_data(F_2017, 2017)

#Combine together and make an ID Variable
F_NSFG             <- rbind(F_2013_vars, F_2015_vars, F_2017_vars)

#Get rid of some of these that we don't need
rm(F_2013, F_2015, F_2017, F_2013_vars, F_2015_vars, F_2017_vars)
```

## Sexuality Definitions 

```{r sexual_behavior}

#Male Partners Ever

#Remove the "Refused" & "Don't Know"
  F_NSFG$num_m_ever[
    F_NSFG$num_m_ever==998 | 
    F_NSFG$num_m_ever==999 ]    <- NA
  
  #Replace number=0 if never had sex 
  F_NSFG$num_m_ever[
    F_NSFG$eversex=="NO"]       <- 0

#Male Partners Past Year
  
  #Remove the "Refused" & "Don't Know" for the year variable
  F_NSFG$num_m_yr[
    F_NSFG$num_m_yr==998 | 
    F_NSFG$num_m_yr==999 |
    F_NSFG$num_m_yr==997 ]      <- NA
  
  #Replace number=0 if never had sex 
  F_NSFG$num_m_yr[
    F_NSFG$eversex=="NO"]       <- 0

#Female Partners Ever

  #Removed "refused" and "don't know" 
  F_NSFG$num_f_ever[
    F_NSFG$num_f_ever==998 | 
    F_NSFG$num_f_ever==999 |
    F_NSFG$num_f_ever==997 ]    <- NA
  
  
  #Replace number=0 if never had sex 
  F_NSFG$num_f_ever[
    F_NSFG$ever_f=="NO"]        <- 0


#Female Partners Past Year

  #Removed "refused" and "don't know" 
  F_NSFG$num_f_yr[
    F_NSFG$num_f_yr==998 | 
    F_NSFG$num_f_yr==999 |
    F_NSFG$num_f_yr==997 ]    <- NA
  
  
  #Replace number=0 if never had sex 
  F_NSFG$num_f_yr[
    F_NSFG$ever_f=="NO"]        <- 0

```

We then pull in and incorporate the 6 year survey weights from NSFG to reflect the US population 

```{r survey_weights}
# # # # Survey Weights  # # # #


weights_6y         <- read.dta13("./2_Inputs/NSFG/weights/weights6y_F.dta")

#Match data from all years with weight file by ID 
F_NSFG             <- merge(F_NSFG, weights_6y, by="CASEID", all = TRUE)


F_NSFG$WGT2011_2017[
  is.na(
F_NSFG$WGT2011_2017
    )]             <- 0



#Create our survey design variable 
nsfg_design        <- svydesign(id      = ~CASEID,
                                strata  = ~strata,
                                weights = ~WGT2011_2017,
                                nest    = TRUE,
                                data    = F_NSFG)

```


Now we define women based on their sexual behavior data, favoring recent data such that: 
```{r table_define1, echo=FALSE}

#Add a formatted table

text_tbl <- data.frame(
  Category = c("Not Sexually Active", "Men Only", "Women Only", "Men and Women"),
  Description = c(
    "No previous sexual partners",
    "Only opposite-sex partners OR majority of all partners or past year partner is opposite-sex with no more than 2 previous same-sex partners",
    "Only same-sex partners OR majority of all partners or past year most recent partner is same-sex with no more than 2 previous opposite-sex partners",
    "1+ same-sex and 1+ opposite-sex partners in the previous year OR
    3+ same-sex and 3+ opposite-sex lifetime partners OR
    3+ same-sex partners lifetime and an opposite-sex partner in the previous year OR
    3+ opposite-sex partners lifetime and a same-sex partner in the previous year"))

kable(text_tbl) %>%
  kable_styling(full_width = F, font="black") %>%
  column_spec(1, bold = T, border_right = T) %>%
  column_spec(2, width = "50em", background = "white") %>%
  row_spec(1, color="#e7298a")  %>%
  row_spec(2, color="#7570b3")  %>%
  row_spec(3, color="#1b9e77")  %>%
  row_spec(4, color="#d95f02")

```

```{r behavior_define}


#Now define bisexual
#First by lifetime behavior
F_NSFG$behav[F_NSFG$num_m_ever>0  & F_NSFG$num_f_ever<3 ]         <- "Men Only" 
F_NSFG$behav[F_NSFG$num_f_ever>0  & F_NSFG$num_m_ever<3 ]         <- "Women Only"

#Then, replace favoring recency (if you haven't had 3 partners of one type we give preference to your more recent behavior)
F_NSFG$behav[F_NSFG$num_f_yr>0]                                   <- "Women Only"
F_NSFG$behav[F_NSFG$num_m_yr>0]                                   <- "Men Only"

#Finally if you haven't had sex with anyone this year and have fewer than 3 of each, 
      #we default to which group you've had more sex with, with ties going to lesbians
F_NSFG$behav[F_NSFG$num_f_yr==0  & F_NSFG$num_m_yr==0 &
                  F_NSFG$num_m_ever<3 & F_NSFG$num_m_ever<3 & 
                    (F_NSFG$num_f_ever>=F_NSFG$num_m_ever)]       <- "Women Only"

#Then change identify as bisexual if you've had three of both or one of each in the previous year
F_NSFG$behav[(F_NSFG$num_f_ever>=3 & F_NSFG$num_m_ever>=3) | 
                  (F_NSFG$num_f_yr>=1   & F_NSFG$num_m_ever>=3) | 
                  (F_NSFG$num_f_ever>=3 & F_NSFG$num_m_yr>=1)  ]  <- "Men and Women"

F_NSFG$behav[F_NSFG$num_f_yr>=1     & F_NSFG$num_m_yr>=1  ]       <- "Men and Women"
#If you've never had sex you are in the none category
F_NSFG$behav[F_NSFG$num_f_ever==0 & F_NSFG$num_m_ever==0 ]        <- "none"

F_NSFG$behav[F_NSFG$sample==0]                                    <- NA

#Update  our survey design variable 
nsfg_design        <- svydesign(id      = ~CASEID,
                                strata  = ~strata,
                                weights = ~WGT2011_2017,
                                nest    = TRUE,
                                data    = F_NSFG)
#Report this with survey weights
  ((svymean(~as.factor(behav),  nsfg_design, na.rm=TRUE))*100) %>%
  kable(digits=(c(2,3))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
    row_spec(4, color="#e7298a")  %>%
    row_spec(2, color="#7570b3")  %>%
    row_spec(3, color="#1b9e77")  %>%
    row_spec(1, color="#d95f02") %>%
  add_header_above(c("Weighted Percent: Sexual Behavior"=3))


```
We also compare by self-reported orientation 
```{r define_orientation , echo=FALSE}
 text_tbl <- data.frame(
   Category = c("Heterosexual", "Homosexual", "Bisexual"),
   Description_Alternative = c(
     "Identifies as straight/heterosexual",
     "Identifies as lesbian/homosexual ",
     "Identifies as bisexual"))
 
 kable(text_tbl) %>%
   kable_styling(full_width = F, font="black") %>%
   column_spec(1, bold = T, border_right = T) %>%
   column_spec(2, width = "50em", background = "white") %>%
   row_spec(1, color="#7570b3")  %>%
   row_spec(2, color="#1b9e77")  %>%
   row_spec(3, color="#d95f02")
```


```{r orientation}
#Pre2016
  F_NSFG$self_orient[F_NSFG$orientation=="Heterosexual or straight"]     <-"heterosexual" 
  F_NSFG$self_orient[F_NSFG$orientation=="Homosexual, gay, or lesbian"]  <-"homosexual"
  F_NSFG$self_orient[F_NSFG$orientation=="Bisexual"]                     <-"bisexual"  
  
  #Create our survey design variable 
nsfg_design        <- svydesign(id      = ~CASEID,
                                strata  = ~strata,
                                weights = ~WGT2011_2017,
                                nest    = TRUE,
                                data    = F_NSFG)
  
     ((svymean(~as.factor(self_orient),  nsfg_design, na.rm=TRUE))*100) %>%
     kable(digits=c(2,3)) %>%
     kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
     column_spec(1, bold = T) %>%
        row_spec(3, color="#1b9e77")  %>%
       row_spec(2, color="#7570b3")  %>%
        row_spec(1, color="#d95f02")  %>%
     add_header_above(c("Raw Percent: Sexual Orientation"=3))
    

    

```

### Sexual Behavior Graphs

We can make graphs to show how each of our behavior categories (in our primary definition) split out by what gender their partners have been in the previous year and in their lifetime. By definition, we don't see any "heterosexual" women who have had sex with a woman in the previous year or more than two women ever. The same goes for lesbian women having no sex with a man in the previous year and sex with not more than two men in their lifetime. Women who are not sexually active are, as we would expect, at 0 for both. 

First let's compare self-report to how we define people

```{r compare_self}
   

prop_behav <- svyby(~self_orient, by=~behav, nsfg_design, 
             svymean, na.rm=TRUE)  

prop_behav1 <- prop_behav[1:5]

rownames(prop_behav1)<- c("Men and Women", "Men Only", "None","Women Only")


 (prop.table(table(F_NSFG$behav, F_NSFG$self_orient), margin = 1)*100) %>%
        kable(digits=c(1,1,1,1), escape = F) %>%
           kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
          column_spec(1, bold = T) %>%

          row_spec(1, color="#d95f02", bold=T)  %>%
          row_spec(2, color="#7570b3", bold=T)  %>%
          row_spec(3, color="#e7298a", bold=T)  %>%
          row_spec(4, color="#1b9e77", bold=T)  %>%
  
  
         column_spec(2, background="#fee0c8")  %>%
         column_spec(3, background="#e4e3f0")  %>%
         column_spec(4, background="#d4f8ed")  %>%

     add_header_above(c("Sexual Orientation vs Sexual Behavior"=4))


```

```{r sex_graphs_behavior, echo=FALSE, warning=FALSE}

#Make Graphs to show how this splits up by ever

ggplot(F_NSFG, aes(x=num_f_ever, y=num_m_ever, color=behav)) +
 scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("Men Only", "Women Only", "Men and Women", "none"), 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "none"="#e7298a", "Men Only"="#7570b3"), na.translate=FALSE) +
  geom_jitter(width=0.3, height=0.3) + theme_classic() +
  xlim(0,40) + ylim(0,50) + ggtitle("Women's Sexual Partners by Partner Gender: Lifetime") +
  ylab("lifetime M sex partners") + xlab("lifetime F sex partners")

#Make Graphs to show how this splits up by past year
ggplot(F_NSFG, aes(x=num_f_yr, y=num_m_yr, color=behav)) +
  scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("Men Only", "Women Only", "Men and Women", "none"), 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "none"="#e7298a", "Men Only"="#7570b3"),
                     na.translate=FALSE) +
  geom_jitter(width=0.3, height=0.3) + theme_classic() +
  xlim(0,8) + ylim(0,15) + ggtitle("Women's Sexual Partners by Partner Gender: Previous Year") +
  ylab("past year M sex partners") + xlab("past year F sex partners")

F_NSFG$partners    <- F_NSFG$num_f_ever + F_NSFG$num_m_ever

#Top code at 50 
F_NSFG$partners[
  F_NSFG$partners>50
  ]                <- 50


#by year
F_NSFG$partners_y  <- F_NSFG$num_f_yr   + F_NSFG$num_m_yr

#Top code at 10
F_NSFG$partners_y[
  F_NSFG$partners_y>4
  ]                <- 4


#ggplot(F_NSFG, aes(x=age, y=partners_y, color=behav)) +
# scale_color_manual(name="Sexual Behavior Group", 
#                     breaks=c("Men Only", "homosexual", "Men and Women", "none"), 
#                     values=c("Men and Women"="#d95f02", "homosexual"="#1b9e77", "none"="#e7298a", "Men Only"="#7570b3"), na.translate=FALSE) +
#  geom_jitter(width=0.3, height=0.3, alpha=0.03) + 
#  theme_classic() +
#   ylim(0,10) + 
#   ggtitle("Number of Partners per Year by Age") +
#  ylab("past year sex partners") + xlab("Age") +
#    stat_summary(fun.data=mean_cl_normal) +
#    geom_smooth(method='lm', se=FALSE)
#

```

##Validation Stats 1: Past year & lifetime partners

Report # of partners ever and in past year by age group. 

```{r age_groups}



#Create age groups 
F_NSFG$age_group[F_NSFG$age>=15 & F_NSFG$age<=19]   <- "15-19"
F_NSFG$age_group[F_NSFG$age>=20 & F_NSFG$age<=24]   <- "20-24"
F_NSFG$age_group[F_NSFG$age>=25 & F_NSFG$age<=29]   <- "25-29"
F_NSFG$age_group[F_NSFG$age>=30 & F_NSFG$age<=34]   <- "30-34"
F_NSFG$age_group[F_NSFG$age>=35 & F_NSFG$age<=39]   <- "35-39"
F_NSFG$age_group[F_NSFG$age>=40 & F_NSFG$age<=50]   <- "40-50"

F_NSFG$partners_y  <- F_NSFG$num_f_yr   + F_NSFG$num_m_yr

#Top code at 15
F_NSFG$partners_y[
  F_NSFG$partners_y>15
  ]                <- 15

#Top code at 15
F_NSFG$num_f_yr[
  F_NSFG$num_f_yr>15
  ]                <- 15

#Top code at 15
F_NSFG$num_m_yr[
  F_NSFG$num_m_yr>15
  ]                <- 15



  
#Update  our survey design variable 
nsfg_design        <- svydesign(id      = ~CASEID,
                                strata  = ~strata,
                                weights = ~WGT2011_2017,
                                nest    = TRUE,
                                data    = F_NSFG)

  

#   
# behav_by_age <- svyby(~behav, by=~age_group, nsfg_design, svymean, na.rm=TRUE)
# colnames(behav_by_age)<- c("Age", "Men and Women", "Men Only", "None", "Homosexual")
#    (behav_by_age[,2:5]*100) %>%  
#     kable(digits=c(2,2,2,2)) %>%
#    kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
#    column_spec(1, bold = T) %>%
#    column_spec(2, color="#d95f02")  %>%
#    column_spec(3, color="#7570b3")  %>%
#    column_spec(4, color="#e7298a")  %>%
#    column_spec(5, color="#1b9e77")  %>%
#    add_header_above(c("Sexual Behavior Category by Age"=5))
  
```

```{r by_behavior_partners}
#Report this with survey weights

  het_byage        <- svyby(~partners_y + num_m_yr + num_f_yr + partners + num_m_ever + num_f_ever,  by=~age_group, 
                            subset(nsfg_design, behav=="Men Only"), 
                     svymean, na.rm=TRUE, vartype =c("ci")) 

  het_byage[1:5]%>%
    kable(digits=(c(2,3))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:6, color="#7570b3")


  hom_byage        <- svyby(~partners_y + num_m_yr + num_f_yr + partners  + num_m_ever + num_f_ever,  by=~age_group, 
                            subset(nsfg_design, behav=="Women Only"),   
                            svymean, na.rm=TRUE, vartype =c("ci")) 
  
  hom_byage[1:5]%>%
    kable(digits=(c(2,3))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:6, color="#1b9e77")

   bi_byage        <- svyby(~partners_y + num_m_yr + num_f_yr + partners  + num_m_ever + num_f_ever,  by=~age_group, 
                            subset(nsfg_design, behav=="Men and Women"),    
                            svymean, na.rm=TRUE, vartype =c("ci")) 
   
   bi_byage[1:5]%>%
    kable(digits=(c(2,3))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:6, color="#d95f02")
   
het_byage$behav    <- "Men Only"
hom_byage$behav    <- "Women Only"
 bi_byage$behav    <- "Men and Women"
 
 byage<- rbind(het_byage,hom_byage, bi_byage)

 
 #Partners in Past Year by Age & Sexual Behavior Group 
(p_behavior <-ggplot(byage, aes(x=age_group, y=partners, col=behav)) +
 geom_point()+
 geom_path(aes(group=behav))+
 geom_pointrange(aes(x=age_group, ymin=ci_l.partners, ymax=ci_u.partners, col=behav))+
 theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  ylim(0, 40) +
  scale_color_manual(name="Sexual Partners", 
                     breaks=c("Men Only", "Women Only", "Men and Women"), 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "Men Only"="#7570b3"),
                     na.translate=FALSE) +
  ggtitle("Mean Number of Sexual Partners: Behavior") +
  ylab("Number of Partners") + xlab("Age Group"))
 
 
#Partners in Past Year by Age & Sexual Behavior Group 
(py_behavior <-ggplot(byage, aes(x=age_group, y=partners_y, col=behav)) +
 geom_point()+
 geom_path(aes(group=behav))+
 geom_pointrange(aes(x=age_group, ymin=ci_l.partners_y, ymax=ci_u.partners_y, col=behav))+
 theme_classic() +
  ylim(0, 6) +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  scale_color_manual(name="Sexual Partners", 
                     breaks=c("Men Only", "Women Only", "Men and Women"), 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "Men Only"="#7570b3"),
                     na.translate=FALSE) +
  ggtitle("Mean Number of Sexual Partners per Year: Behavior") +
  ylab("Number of Partners") + xlab("Age Group"))



#M and F Partners in Past Year among bisexuals 
(bi_graph <-  ggplot(bi_byage) +
 geom_path(aes(x=age_group, y=num_m_yr, group=behav), linetype="dashed", col="#7570b3") +
 geom_pointrange(aes(x=age_group, y=num_m_yr, ymin=ci_l.num_m_yr, ymax=ci_u.num_m_yr), col="#7570b3") +
  geom_path(aes(x=age_group, y=num_f_yr, group=behav), linetype="dotted", col="#1b9e77") +
 geom_pointrange(aes(x=age_group, y=num_f_yr, ymin=ci_l.num_f_yr, ymax=ci_u.num_f_yr), col="#1b9e77") +  
 annotate(geom="text", x=1.5, y=0.9, label="Female Partners", color="#1b9e77", size=2.5) +
 annotate(geom="text", x=1.5, y=2.5, label="Male Partners", color="#7570b3", size=2.5) +
     theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
    ylim(0, 3) +
  scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("New Male Partners", "New Female Partners"), 
                     na.translate=FALSE) +
  ggtitle("Mean Number of M/F Partners per Year 
          (Bisexuals- Behavior)") +
  ylab("Number of Partners") + xlab("Age Group"))

#Save this - first create a workbook 
wb_ts                <- createWorkbook()
 
addWorksheet(wb_ts, "validation_out_F")

writeData(wb_ts, sheet= "validation_out_F", byage)

```



Check how this compares to when we use self-reported orientation: 

```{r bisexual_self}

  orient_by_age <- svyby(~self_orient, by=~age_group, nsfg_design, svymean, na.rm=TRUE)
  colnames(orient_by_age)<- c("Age", "Bisexual", "Heterosexual", "Homosexual")

     (orient_by_age[,2:4]*100) %>%  
      kable(digits=c(2,2,2,2)) %>%
     kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
     column_spec(1, bold = T) %>%
     column_spec(4, color="#1b9e77")  %>%
     column_spec(3, color="#7570b3")  %>%
     column_spec(2, color="#d95f02")  %>%
     add_header_above(c("Self-Reported Sexual Orientation by Age"=4))
```

```{r by_orientation, echo=FALSE}
#Report this with survey weights

  het_byage2 <- svyby(~partners_y + num_m_yr + num_f_yr + partners ,  by=~age_group, 
                      subset(nsfg_design, self_orient=="heterosexual" & eversex=="YES"), 
                      svymean, na.rm=TRUE, vartype =c("ci")) 

#het_byage2[1:5]%>%
#  kable(digits=(c(2,3))) %>%
#kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
#column_spec(1, bold = T) %>%
#column_spec(2:6, color="#7570b3")


  hom_byage2 <- svyby(~partners_y + num_m_yr + num_f_yr + partners ,  by=~age_group, 
                      subset(nsfg_design, self_orient=="homosexual" & eversex=="YES"),   
                      svymean, na.rm=TRUE, vartype =c("ci")) 
  
#  hom_byage2[1:5]%>%
#    kable(digits=(c(2,3))) %>%
#  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
#  column_spec(1, bold = T) %>%
#  column_spec(2:6, color="#1b9e77")

   bi_byage2 <- svyby(~partners_y + num_m_yr + num_f_yr + partners ,  by=~age_group, 
                      subset(nsfg_design, self_orient=="bisexual" & eversex=="YES"),     
                      svymean, na.rm=TRUE, vartype =c("ci")) 
   
#   bi_byage2[1:5]%>%
#    kable(digits=(c(2,3))) %>%
#  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
#  column_spec(1, bold = T) %>%
#  column_spec(2:6, color="#d95f02")
   
het_byage2$self_orient    <- "heterosexual"
hom_byage2$self_orient    <- "homosexual"
 bi_byage2$self_orient    <- "bisexual"
 
 byage2 <- rbind(het_byage2,hom_byage2, bi_byage2)
 

 #Partners in Past Year by Age & Sexual Behavior Group 
(p_orientation <- ggplot(byage2, aes(x=age_group, y=partners, col=self_orient)) +
 geom_point()+
 geom_path(aes(group=self_orient))+
 geom_pointrange(aes(x=age_group, ymin=ci_l.partners, ymax=ci_u.partners, col=self_orient))+
 theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  ylim(0, 30) +
  scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("heterosexual", "homosexual", "bisexual"), 
                     values=c("bisexual"="#d95f02", "homosexual"="#1b9e77", "heterosexual"="#7570b3"),
                     na.translate=FALSE) +
  ggtitle("Mean Number of Sexual Partners: Orientation") +
  ylab("Number of Partners") + xlab("Age Group"))

 
  
#Partners in Past Year by Age & Sexual Behavior Group 
(py_orientation <- ggplot(byage2, aes(x=age_group, y=partners_y, col=self_orient)) +
 geom_point()+
 geom_path(aes(group=self_orient))+
 geom_pointrange(aes(x=age_group, ymin=ci_l.partners_y, ymax=ci_u.partners_y, col=self_orient))+
 theme_classic() +
     ylim(0, 5) +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("heterosexual", "homosexual", "bisexual"), 
                     values=c("bisexual"="#d95f02", "homosexual"="#1b9e77", "heterosexual"="#7570b3"),
                     na.translate=FALSE) +
  ggtitle("Mean Number of Partners per Year: Orientation") +
  ylab("Number of Partners") + xlab("Age Group"))



#M and F Partners in Past Year among bisexuals 
(bi_graph2 <-  ggplot(bi_byage2) +
 geom_path(aes(x=age_group, y=num_m_yr, group=self_orient), linetype="dashed", col="#7570b3") +
 geom_pointrange(aes(x=age_group, y=num_m_yr, ymin=ci_l.num_m_yr, ymax=ci_u.num_m_yr), col="#7570b3") +
  geom_path(aes(x=age_group, y=num_f_yr, group=self_orient), linetype="dotted", col="#1b9e77") +
 geom_pointrange(aes(x=age_group, y=num_f_yr, ymin=ci_l.num_f_yr, ymax=ci_u.num_f_yr), col="#1b9e77") +  
 annotate(geom="text", x=1.5, y=0.2, label="Female Partners", color="#1b9e77", size=2.5) +
 annotate(geom="text", x=1.5, y=2.0, label="Male Partners", color="#7570b3", size=2.5) +
     theme_classic() +
  theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  ylim(0, 2.5) +
  ggtitle("Mean Number of M/F Partners per Year
          (Bisexuals- Orientation)") +
  ylab("Number of Partners") + xlab("Age Group"))





```

Combine and Compare

```{r combine, fig.height =10, fig.width = 8, fig.align = "center"}

p_women <- ggarrange(p_behavior, p_orientation, py_behavior, py_orientation, bi_graph, bi_graph2, 
                    labels = c("A1", "A2", "B1", "B2", "C1", "C2"),
                    ncol = 2, 
                    nrow = 3, 
                    common.legend=TRUE, 
                    legend="bottom")

(p_women <-   annotate_figure(p_women,
               top = text_grob("Sexual Partner Change Among Sexually Active Women", color = "Black", face = "bold", size = 14)))
               
ggexport(p_women, width=8, height=10, filename="./compare_NSFG_F.pdf")

```

```{r age_by}

#correct these to remove any missing
F_NSFG$age_first[F_NSFG$age_first>90]  <- NA
F_NSFG$age_same[F_NSFG$age_same>90]   <- NA

#Bottom-code age, no one can have sex before 10 years of age in my sample
F_NSFG$age_first[F_NSFG$age_first<10]  <- 10
F_NSFG$age_same[F_NSFG$age_same<10]   <- 10

F_NSFG$age_first_behav                                <- 99

#Create a new variable that divides this by behavior
F_NSFG$age_first_behav[
  F_NSFG$behav=="Women Only" 
         & !is.na(F_NSFG$behav)]       <- F_NSFG$age_same[F_NSFG$behav=="Women Only" & !is.na(F_NSFG$behav)]
F_NSFG$age_first_behav[
  F_NSFG$behav=="Men Only" 
          & !is.na(F_NSFG$behav)]      <- F_NSFG$age_first[F_NSFG$behav=="Men Only"  & !is.na(F_NSFG$behav)] 

#Pick which is earlier for bisexual individuals 
F_NSFG$age_lower                       <- pmin(F_NSFG$age_first, F_NSFG$age_same)

F_NSFG$age_first_behav[
  F_NSFG$behav=="Men and Women"
            & !is.na(F_NSFG$behav)]    <- F_NSFG$age_lower[ F_NSFG$behav=="Men and Women"  & !is.na(F_NSFG$behav)]



  #Create our survey design variable 
nsfg_design        <- svydesign(id      = ~CASEID,
                                strata  = ~strata,
                                weights = ~WGT2011_2017,
                                nest    = TRUE,
                                data    = F_NSFG)

 first_by_behav <- svyby(~age_first_behav ,  by=~behav, 
                      subset(nsfg_design, eversex=="YES"), 
                      svymean, na.rm=TRUE, vartype =c("ci")) 


 first_by_behav <- svyby(~age_first_behav ,  by=~behav, 
                      subset(nsfg_design, eversex=="YES" & age_first<90), 
                      svymean, na.rm=TRUE, vartype =c("ci")) 

(first_w      <- 
  ggplot() + 
  geom_bar(aes(y=first_by_behav$age_first, x=first_by_behav$behav, fill=first_by_behav$behav), 
           stat="identity",  color="black") +
           ylab("Age at first sexual experience") + xlab(NULL) +
   geom_errorbar(aes(x=first_by_behav$behav, ymin=first_by_behav$ci_l, ymax=first_by_behav$ci_u), width=.05) +
            scale_fill_manual(name=NULL, 
                     values=c("Men and Women"="#fd862a", "Women Only"="#2bdba6", "none"="#ef6eb0", "Men Only"="#a5a2ce")) +
               theme(legend.position=NULL, 
         legend.box = "vertical", 
         panel.background = element_rect(fill = NA),) + 
         ggtitle("Age at First Sex: Women by Sexual Behavior Category"))



(first_w2      <- 
  ggplot() + 
  geom_point(aes(y=first_by_behav$age_first, x=first_by_behav$behav, color=first_by_behav$behav), 
           stat="identity",  size=2) +
           ylab("Age at first sexual experience") + xlab(NULL) +
   geom_errorbar(aes(x=first_by_behav$behav, ymin=first_by_behav$ci_l, ymax=first_by_behav$ci_u , color=first_by_behav$behav), width=.1, size=1) +
    ylim(12, 22) +
            scale_color_manual(name=NULL, 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "Men Only"="#7570b3")) +
               theme(legend.position=NULL, 
         legend.box = "vertical", 
         panel.background = element_rect(fill = NA),) + 
         ggtitle("Age at First Sex: Women by Sexual Behavior Category"))

```

##Target Stats 1- Duration

  We use the variable "PCURRNT" which reports whether you identify your most recent partner as a current partner. For those who have only ever had one sexual partner, they did not get asked this question, so we instead identify whether their first partner is still an ongoing sexual partner. 
  
  We then use the computed *duration* variable that gives us the length of the relationship. We restrict to only current relationships using report of each. (We check our own count against num_curr and they match)
  
  We then also create a variable that indicates the duration by mean degree (are people in multiple concurrent relationships in shorter relationships on average). For those in multiple relationships, we average the length of all ongoing relationships. 
  
  
```{r duration1}

#First correct the changing capitalization from year to year 

F_NSFG$current_p_recent[
  F_NSFG$current_p_recent=="YES"
  ]                 <- "Yes"

F_NSFG$current_p_recent[
  F_NSFG$current_p_recent=="NO"
  ]                <- "No"


#Then, if you have only ever had one partner and you stated that your first partner is a current partner, 
      # we identify you as currently in a partnership

#F_NSFG$current_p_recent[
#  F_NSFG$num_m_ever==1 &
#  F_NSFG$current_p_first=="Yes"
#  ]                <-"Yes"


F_NSFG$current_p_recent[
  F_NSFG$num_m_ever==1 &
  F_NSFG$num_curr==1
  ]                <-"Yes"


# # # # # # # # # # Duration # # # # # # # # # # 

#One time partnerships are labeled as 997- I want to make a new variable to capture these and then set them to NA
F_NSFG$lastp_onetime         <- 
F_NSFG$twop_onetime          <- 
F_NSFG$threep_onetime        <- 
F_NSFG$count                 <- 0 

# # # # # Most recent partnership # # # # # 

F_NSFG$lastp_onetime[
  F_NSFG$duration_1==997]    <- 1

#Replace the 997s
F_NSFG$duration_1[
  F_NSFG$duration_1==997]    <- NA


#Drop anyone who claims to have started a sexual relationship before 10 years of age 
F_NSFG$duration_1[
  (F_NSFG$age-(F_NSFG$duration_1/12))<=10
  ]                          <- NA


#Create new variable which we'll change to define only current relationships 
F_NSFG$duration_1c           <- F_NSFG$duration_1


#Remove if we don't know whether this is a current relationship
F_NSFG$duration_1c[
  is.na(F_NSFG$current_p_recent)
  ]                          <-NA     


#Remove if it is defintely not a current relationship
F_NSFG$duration_1c[
  F_NSFG$current_p_recent=="No" & F_NSFG$current_p_first!="Yes"
  ]                          <-NA     


#Also drop anyone who is not in any current sexual relationships
F_NSFG$duration_1c[
  F_NSFG$num_crrr==0]        <- NA

F_NSFG$current_p_recent[is.na(F_NSFG$current_p_recent)]        <-  "No"


F_NSFG$count[F_NSFG$current_p_recent=="Yes"]        <-  F_NSFG$count[F_NSFG$current_p_recent=="Yes"] + 1 

```

```{r duration2, echo=FALSE}
## # # # # Second partnership # # # # # 

F_NSFG$twop_onetime[
  F_NSFG$duration_2==997]    <- 1

#Replace the 997s
F_NSFG$duration_2[
  F_NSFG$duration_2==997]    <- NA


#Drop anyone who claims to have started a sexual relationship before 10 years of age 
F_NSFG$duration_2[
  (F_NSFG$age-(F_NSFG$duration_2/12))<=10
  ]                          <- NA


#Create new variable which we'll change to define only current relationships 
F_NSFG$duration_2c            <- F_NSFG$duration_2
#


F_NSFG$current_p2[
  F_NSFG$current_p2=="YES"
  ]                 <- "Yes"

F_NSFG$current_p2[
  F_NSFG$current_p2=="NO"
  ]                <- "No"
  
#Remove if we don't know whether this is a current relationship
F_NSFG$duration_2c[
  is.na(F_NSFG$current_p2)
  ]                          <-NA     

#Remove if it is defintely not a current relationship
F_NSFG$duration_2c[
  F_NSFG$current_p2=="No"
  ]                          <-NA    

F_NSFG$current_p2[is.na(F_NSFG$current_p2)]     <-  "No"
F_NSFG$count[F_NSFG$current_p2=="Yes"]          <-  F_NSFG$count[F_NSFG$current_p2=="Yes"] + 1 

```

```{r duration3}
## # # # # THird partnership # # # # # 

F_NSFG$threep_onetime[
  F_NSFG$duration_3==997]    <- 1

#Replace the 997s
F_NSFG$duration_3[
  F_NSFG$duration_3==997]    <- NA


#Drop anyone who claims to have started a sexual relationship before 10 years of age 
F_NSFG$duration_3[
  (F_NSFG$age-(F_NSFG$duration_3/12))<=10
  ]                          <- NA


#Create new variable which we'll change to define only current relationships 
F_NSFG$duration_3c            <- F_NSFG$duration_3
#


F_NSFG$current_p3[
  F_NSFG$current_p3=="YES"
  ]                 <- "Yes"

F_NSFG$current_p3[
  F_NSFG$current_p3=="NO"
  ]                <- "No"
  
#Remove if we don't know whether this is a current relationship
F_NSFG$duration_3c[
  is.na(F_NSFG$current_p3)
  ]                          <-NA     

#Remove if it is defintely not a current relationship
F_NSFG$duration_3c[
  F_NSFG$current_p3=="No"
  ]                          <-NA    


F_NSFG$current_p3[is.na(F_NSFG$current_p3)]     <-  "No"
F_NSFG$count[F_NSFG$current_p3=="Yes"]          <-  F_NSFG$count[F_NSFG$current_p3=="Yes"] + 1 


# # # # # # # # # # Duration by Mean Degree # # # # # # # # # # # # # # # # # # # # 
F_NSFG$duration_1c[
  is.na(F_NSFG$duration_1c)]          <-0 
F_NSFG$duration_2c[
  is.na(F_NSFG$duration_2c)]          <-0 
F_NSFG$duration_3c[
  is.na(F_NSFG$duration_3c)]          <-0   


F_NSFG$avg_dur[
  F_NSFG$num_curr==0]                 <- 0

F_NSFG$avg_dur[F_NSFG$num_curr>0]            <- (F_NSFG$duration_1c[F_NSFG$num_curr>0]  + 
                                                 F_NSFG$duration_2c[F_NSFG$num_curr>0]   + 
                                                 F_NSFG$duration_3c[F_NSFG$num_curr>0] ) / F_NSFG$count[F_NSFG$num_curr>0]

ggplot() + 
    geom_point(aes(y=F_NSFG$avg_dur/12, x=F_NSFG$age, color=F_NSFG$num_curr))

table(F_NSFG$count, F_NSFG$num_curr)

#We are going to combine 2 and 3
 F_NSFG$num_curr[F_NSFG$num_curr==3]         <- 2

#Update  our survey design variable 
nsfg_design                  <- svydesign(id      = ~CASEID,
                                         strata  = ~strata,
                                         weights = ~WGT2011_2017,
                                         nest    = TRUE,
                                         data    = F_NSFG)

```

Graph of duration by sexual behavior & concurrency

```{r duration_M_by_behav}

het_duration             <- svyby(~duration_1c, by=~~num_curr + age_group  , 
                              subset(nsfg_design, behav=="Men Only"),
                              svymean, na.rm=TRUE, vartype =c("ci"))
      
het_duration[1:4]%>%
  kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:5, color="#7570b3")
                      
                      
bi_duration             <- svyby(~duration_1c,  by=~~num_curr + age_group , 
                            subset(nsfg_design, behav=="Men and Women"),
                            svymean, na.rm=TRUE, vartype =c("ci")) 
                  
bi_duration[1:4]%>%
  kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:5, color="#d95f02")
   

het_duration$behav      <- "Men Only"
bi_duration$behav       <- "Men and Women"

 
 byage_dur              <- rbind(het_duration,bi_duration)

  byage_dur1 <-    byage_dur[which(byage_dur$num_curr==1),]                
  byage_dur2 <-    byage_dur[which(byage_dur$num_curr==2),]                

#Partners in Past Year by Age & Sexual Behavior Group 
 (d_behavior <- ggplot(byage_dur1, aes(x=age_group, 
                                     y=duration_1c/12, 
                                     col=behav)) +
 geom_point()  +
 geom_path(aes(group=behav))+
 geom_pointrange(aes(x=age_group, 
                       ymin=ci_l/12, 
                      ymax=ci_u/12, col=behav)) +
      theme_classic() +
      theme(panel.background = element_rect(fill = NA),
      plot.title = element_text(size = 10, face = "bold")) +
      scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("Men Only", "Men and Women"), 
                     values=c("Men Only"="#7570b3", "Men and Women"="#d95f02"),
                     na.translate=FALSE) +
       ggtitle("Mean Relationship Duration by Sexual Behavior (1 Partner)") +
       ylab("Relationship Duration (Years)") + 
       xlab("Age Group"))
  
  
  (d_behavior <- 
      ggplot(byage_dur2, 
             aes(x=age_group, y=duration_1c/12, col=behav)) +
               geom_point()  +
               geom_path(aes(group=behav))+
               geom_pointrange(aes(x=age_group, 
                                     ymin=ci_l/12, 
                                    ymax=ci_u/12, col=behav)) +
                    theme_classic() +
                    theme(panel.background = element_rect(fill = NA),
                    plot.title = element_text(size = 10, face = "bold")) +
                    scale_color_manual(name="Sexual Behavior Group", 
                                   breaks=c("Men Only", "Men and Women"), 
                                   values=c("Men Only"="#7570b3", "Men and Women"="#d95f02"),
                                   na.translate=FALSE) +
                     ggtitle("Mean Relationship Duration by Sexual Behavior (2+ Partners)") +
                     ylab("Relationship Duration (Years)") + 
                     xlab("Age Group"))
             
 
 #Partners in Past Year by Age & Sexual Behavior Group 
#(d_behavior <- ggplot(byage_dur, aes(x=byage_dur$age_group[byage_dur$num_curr==1], 
#                                     y=byage_dur$duration_1c[byage_dur$num_curr==1]/12, 
#                                     col=byage_dur$behav[byage_dur$num_curr==1])) +
# geom_point()  +
# geom_path(aes(group=behav))+
# geom_pointrange(aes(x=age_group[byage_dur$num_curr==1], 
#                       ymin=ci_l[byage_dur$num_curr==1]/12, 
#                       ymax=ci_u[byage_dur$num_curr==1]/12, col=behav)) +
#      theme_classic() +
#      theme(panel.background = element_rect(fill = NA),
#      plot.title = element_text(size = 10, face = "bold")) +
#      scale_color_manual(name="Sexual Behavior Group", 
#                     breaks=c("Men Only", "Men and Women"), 
#                     values=c("Men Only"="#7570b3", "Men and Women"="#d95f02"),
#                     na.translate=FALSE) +
#       ggtitle("Mean Relationship Duration by Sexual Behavior") +
#       ylab("Relationship Duration (Years)") + 
#       xlab("Age Group"))
#
het_duration             <- svyby(~duration_1c, by=~~num_curr  , 
                              subset(nsfg_design, behav=="Men Only"),
                              svymean, na.rm=TRUE, vartype =c("ci"))
    
                      
bi_duration             <- svyby(~duration_1c,  by=~~num_curr  , 
                            subset(nsfg_design, behav=="Men and Women"),
                            svymean, na.rm=TRUE, vartype =c("ci"))   
  

addWorksheet(wb_ts, "duration_out_F")

writeData(wb_ts, sheet= "duration_out_F", byage_dur)


```
##Target Stats 2- Concurrency/Mean Degree

Above we identified one time partnerships and created a count variable here we report precent of relationships that are concurrent by sexual behavior. 

```{r concurrency}

het_curr    <- prop.table(svytable(~num_curr + age_group, 
                      subset(nsfg_design, behav=="Men Only")),2)
het_curr%>%
    kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:7, color="#7570b3")
                      
het_curr_df   <- data_frame("O partners"=het_curr[1,], 
                           "1 partner" =het_curr[2,],
                          "2+ partners"=het_curr[3,], 
                            "age group"= c("15-19", "20-24", "25-29", "30-34", "35-39", "40-50"), 
                         "behav"="Men Only")

           
bi_curr    <- prop.table(svytable(~num_curr + age_group, 
                      subset(nsfg_design, behav=="Men and Women")),2)           

bi_curr%>%
    kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:7, color="#d95f02")

bi_curr_df   <- data_frame("O partners"=bi_curr[1,], 
                         "1 partner" =bi_curr[2,],
                         "2+ partners"=bi_curr[3,], 
                         "age group"= c("15-19", "20-24", "25-29", "30-34", "35-39", "40-50"), 
                         "behav"="Men and Women")

curr_all     <- rbind(het_curr_df, bi_curr_df)


addWorksheet(wb_ts, "meandeg_out_F")

writeData(wb_ts, sheet= "meandeg_out_F", curr_all)      

```



```{r annual meandeg}

  F_NSFG$annual_md           <- "0" 
  F_NSFG$annual_md[
    F_NSFG$partners_y==1]    <- "1" 
  F_NSFG$annual_md[
    F_NSFG$partners_y>1]     <- "2+"
  
  
  prop.table(table(F_NSFG$annual_md[F_NSFG$behav=="Men Only"]))
  prop.table(table(F_NSFG$num_curr[F_NSFG$behav=="Men Only"]))
  
  
  conversion_factor          <- prop.table(table(F_NSFG$annual_md[F_NSFG$behav=="Men Only"],
                                                 F_NSFG$num_curr[F_NSFG$behav=="Men Only"]), 1)
  
  
  
  #Update  our survey design variable 
      nsfg_design            <- svydesign(id      = ~CASEID,
                                          strata  = ~strata,
                                          weights = ~WGT2011_2017,
                                          nest    = TRUE,
                                          data    = F_NSFG)
          
    
      t(conversion_factor)%*%  ((as.matrix((svymean(~as.factor(annual_md), 
      subset(nsfg_design, behav=="Women Only"), na.rm=TRUE, estimates.only=TRUE ))))*100) %>%
     kable(digits=c(2,3)) %>%
     kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
     column_spec(1, bold = T) %>%
     row_spec(1:3, color="#1b9e77") 
      
  
  
       t(conversion_factor)%*%((as.matrix((svymean(~as.factor(annual_md), 
      subset(nsfg_design, behav=="Men Only"), na.rm=TRUE, estimates.only=TRUE ))))*100) %>%
     kable(digits=c(2,3)) %>%
     kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
     column_spec(1, bold = T) %>%
     row_spec(1:3, color="#7570b3") 
      
  
  
      t(conversion_factor)%*%((as.matrix((svymean(~as.factor(annual_md), 
      subset(nsfg_design, behav=="Men and Women"), na.rm=TRUE, estimates.only=TRUE ))))*100) %>%
     kable(digits=c(2,3)) %>%
     kable_styling(bootstrap_options = c("condensed", full_width= F, position="center"))%>%
     column_spec(1, bold = T) %>%
     row_spec(1:3, color="#d95f02") 
      
      
    

```

            

##Target Stats 3- One-time partnerships

We get a report of whether participants had any one-time partnerships during the year, and report this by 

```{r one_time}

#table(F_NSFG$twop_onetime)
#table(F_NSFG$threep_onetime)

F_NSFG$any_onetime           <- 0

F_NSFG$any_onetime[
  F_NSFG$lastp_onetime==1 | 
  F_NSFG$twop_onetime==1  | 
  F_NSFG$threep_onetime==1]  <- 1

table(F_NSFG$any_onetime, F_NSFG$lastp_onetime, useNA="ifany")   #247 for most recent
table(F_NSFG$any_onetime, F_NSFG$twop_onetime, useNA="ifany")    #103 for second-most recent
table(F_NSFG$any_onetime, F_NSFG$threep_onetime, useNA="ifany")  #37 for third-most recent 


F_NSFG$num_onetime           <-  (F_NSFG$lastp_onetime +  F_NSFG$twop_onetime +  F_NSFG$threep_onetime)
table(F_NSFG$num_onetime)  #The majority of people only had 1 one-time partners but some had 2 or 3. 

#Update  our survey design variable 
nsfg_design                  <- svydesign(id      = ~CASEID,
                                         strata  = ~strata,
                                         weights = ~WGT2011_2017,
                                         nest    = TRUE,
                                         data    = F_NSFG)

#table(F_NSFG$any_onetime)

#table(F_NSFG$any_onetime, F_NSFG$age_group, F_NSFG$behav)


het_onetime            <- svyby(~any_onetime,  by=~age_group, 
                          subset(nsfg_design, behav=="Men Only"),
                          svymean, na.rm=TRUE, vartype =c("ci"))

het_onetime[1:4]%>%
    kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:5, color="#7570b3")
                      
                      
bi_onetime             <- svyby(~any_onetime,  by=~age_group , 
                          subset(nsfg_design, behav=="Men and Women"),
                          svymean, na.rm=TRUE, vartype =c("ci")) 

  
          
bi_onetime[1:4]%>%
    kable(digits=(c(2))) %>%
  kable_styling(bootstrap_options = c("condensed", full_width= F, position="center")) %>%
  column_spec(1, bold = T) %>%
  column_spec(2:5, color="#d95f02")
   
het_onetime$behav      <- "Men Only"
 bi_onetime$behav      <- "Men and Women"

 byage_onetime         <- rbind(het_onetime,bi_onetime)
 
                      
                      

 #Partners in Past Year by Age & Sexual Behavior Group 
(d_behavior           <- ggplot(byage_onetime, aes(x=age_group, y=any_onetime, col=behav)) +
 geom_point()  +
 geom_path(aes(group=behav))+
   ylim(-0.02,0.5) + 
 geom_pointrange(aes(x=age_group, ymin=ci_l, ymax=ci_u, col=behav)) +
 theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("Men Only", "Men and Women"), 
                     values=c("Men Only"="#7570b3", "Men and Women"="#d95f02"),
                     na.translate=FALSE) +
  ggtitle("Percent with past year one-time partnerships") +
  ylab("Percent with one-time relationship in past year") + xlab("Age Group"))
 
 


 onetime_mean         <- svyby(~any_onetime,  by=~num_curr, 
                          nsfg_design,
                          svymean, na.rm=TRUE, vartype =c("ci")) 
 
  #Partners in Past Year by Age & Sexual Behavior Group 
(d_behavior           <- ggplot(onetime_mean, aes(x=num_curr, y=any_onetime)) +
 geom_point()  +
 geom_path(aes(group=as.factor(num_curr))) +
 geom_pointrange(aes(x=num_curr, ymin=ci_l, ymax=ci_u, col=as.factor(num_curr))) +
 theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
  ggtitle("Percent with past year one-time partnerships") +
  ylab("Percent with one-time relationship in past year") + xlab("Number Current"))


 addWorksheet(wb_ts, "onetime_age_F")
writeData(wb_ts, sheet= "onetime_age_F", byage_onetime)


 addWorksheet(wb_ts, "onetime_degree_F")
writeData(wb_ts, sheet= "onetime_degree_F", onetime_mean)
```

##Target Stats 4: Age Assortiveness

We take the age of your current partner(s) and your own age and we compute the difference. 

```{r age_difference}

F_NSFG$ageat_p_recent[F_NSFG$ageat_p_recent>=97] <- NA
F_NSFG$ageat_p_recent[F_NSFG$ageat_p_recent<=10] <- NA
F_NSFG$page_p_recent[F_NSFG$page_p_recent>=97]  <- NA

ggplot() + 
    geom_jitter(aes(x=F_NSFG$ageat_p_recent, y=F_NSFG$page_p_recent, col=F_NSFG$behav)) + 
    ylab("Partner's Age") + 
    xlab("Respondent's Age") + 
  geom_abline( intercept=0, slope=1, color="gray", linetype="dashed") + 
   theme_classic() +
      theme(panel.background = element_rect(fill = NA),
        plot.title = element_text(size = 10, face = "bold")) +
   scale_color_manual(name="Sexual Behavior Group", 
                     breaks=c("Men and Women", "Women Only", "Men Only", "none"), 
                     values=c("Men and Women"="#d95f02", "Women Only"="#1b9e77", "Men Only"="#7570b3", "none"="gray"),
                     na.translate=FALSE) +
  ggtitle("Comparison of Age of Sexual Partners by Respondent age and sexual behavior")



F_NSFG$abs_diff         <- abs(F_NSFG$ageat_p_recent- F_NSFG$page_p_recent)
F_NSFG$sq_abs_diff      <- abs( sqrt(F_NSFG$ageat_p_recent) - sqrt(F_NSFG$page_p_recent))


#Update  our survey design variable 
nsfg_design                  <- svydesign(id      = ~CASEID,
                                         strata  = ~strata,
                                         weights = ~WGT2011_2017,
                                         nest    = TRUE,
                                         data    = F_NSFG)


# # # # # Abs Diff # # # #

het_absdiff            <- svyby(~abs_diff,  by=~age_group, 
                          subset(nsfg_design, behav=="Men Only"),
                          svymean, na.rm=TRUE, vartype =c("ci"))

bi_absdiff             <- svyby(~abs_diff,  by=~age_group, 
                          subset(nsfg_design, behav=="Men and Women"),
                          svymean, na.rm=TRUE, vartype =c("ci"))

het_absdiff$behav      <- "Men Only"
 bi_absdiff$behav      <- "Men and Women"

 abs_diff_byage        <- rbind(het_absdiff,bi_absdiff)
 
        
  (d_abs_diff <- 
      ggplot(abs_diff_byage, aes(x=age_group, y=abs_diff, col=behav)) +
               geom_point()  +
               geom_path(aes(group=behav))+
               geom_pointrange(aes(x=age_group, 
                                     ymin=ci_l, 
                                    ymax=ci_u, col=behav)) +
                    theme_classic() +
                    theme(panel.background = element_rect(fill = NA),
                    plot.title = element_text(size = 10, face = "bold")) +
                    scale_color_manual(name="Sexual Behavior Group", 
                                   breaks=c("Men Only", "Men and Women"), 
                                   values=c("Men Only"="#7570b3", "Men and Women"="#d95f02"),
                                   na.translate=FALSE) +
                     ggtitle("Age Difference between sexual partners") +
                     ylab("Absolute age difference (Years)") + 
                     xlab("Age Group"))

 
 addWorksheet(wb_ts, "age_assort")
writeData(wb_ts, sheet= "age_assort", abs_diff_byage)


#Save my workbook object to an excel file
saveWorkbook(wb_ts, "./3_Outputs/target_stats_out.xlsx", overwrite = TRUE)
 

```
