==> R Code/08_Combined_Model_Test.Rmd <==
---
title: "08_Combined_Model_Test"
author: "Jenny Spencer"
date: "4/1/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

rm(list=ls())

library(EpiModel)    #the primary package we will be using to make our models 
library(ndtv)        #for the fancy videos
library(readxl)      #read in excel files 
library(tidyverse)   #Includes a number of shortcuts for data manipulation
library(beepr)       #makes the code beep when you get to it. Beep. 
library(kableExtra)  #for making tables look nice
library(mc2d)        #for using pert distribution
#remotes::install_github("statnet/tergmLite@v1.2.0") #the TERGM lite package
library(tergmLite)   #for running the network code
library(sessioninfo)
sessioninfo::package_info(pkgs = c("network", "networkDynamic", "statnet.common",
                                   "ergm", "tergm", "EpiModel", "EpiModelHPC",
                                   "tergmLite", "EpiABC", "EpiModelHIV",
                                   "ARTnetData", "ARTnet"), dependencies = FALSE)

#install.packages("sessioninfo")

```

## Model Data 

  This will eventually pull directly from NSFG files, but hardcoded here for simplicity/portability as we test approaches

```{r data_in}

p.msmo                <- 0.2
p.msmw                <- 0.2
p.mswo                <- 1-(p.msmo + p.msmw)

p.pos_r               <- .16   
p.pos_i               <- .28 
p.pos_v               <- .56 

msm_duration          <- 71.6
msw_duration          <- 89.5

p.f                   <- 0.35

```

### Simulated Population

Using realistic percentages of sexual minority prevalence 

```{r population}
#Create a nework of 50/50 men and women 
n                       <- 10000

# # # Sex # # # 
#define sex
v.sex                   <- c(sample(c("F", "M"), n, 
                                        replace= TRUE, 
                                        prob=c(p.f, (1-p.f)))) 


# # # Sexual Behavior # # # 
#define sexual behavior based on sex
v.behave                <- ifelse(v.sex=="F", "wsmo", c(sample(c("msmo", "msmw", "mswo"), sum(v.sex=="M"), 
                                        replace= TRUE, 
                                        prob=c(p.msmo, p.msmw, p.mswo)))
)

#get these numbers
n.msmo                  <-sum(v.behave=="msmo")
n.msmw                  <-sum(v.behave=="msmw")
n.mswo                  <-sum(v.behave=="mswo")
n.wsmo                  <-sum(v.behave=="wsmo")



#define position preference based on sexual behavior. For now assume positional preference distribution is similar
v.position              <- ifelse((v.behave=="msmo"| v.behave=="msmw"), c(sample(c("Receptive",          
                                                                                   "Insertive", 
                                                                                    "Versatile"), 
                                                                        sum(v.behave=="msmo"| v.behave=="msmw"), 
                                                                        replace= TRUE, 
                                                                         prob=c(p.pos_r, p.pos_i, p.pos_v))), "none")

#Define a variable that dictates what age you begin sexual activity
v.age_active              <- ifelse(v.behave=="msmo", c(rpert(sum(n.msmo), min=8, mode=18, max=25, shape=10)), 
                             ifelse(v.behave=="msmw", c(rpert(sum(n.msmw), min=8, mode=14, max=25, shape=10)),
                             ifelse(v.behave=="mswo", c(rpert(sum(n.mswo), min=8, mode=17, max=25, shape=10)),      
                                                     (c(rpert(sum(n.wsmo), min=8, mode=18, max=25, shape=10))))))

# # #  AGE  # # # 

#calc age dist 
l.age                   <- c( "<1", "1-4", "5-9", "10-14", "15-20", "20-24", "25-29", "30-34", "35-39", "40-44", "45-49", "50-54", "55-59",   
                         "60-64", "65-69", "70-74", "75-79",  "80-85", "85+")

#make upper and lower limit variables- the actual age will be draw from a distribution within these so these will line up exactly. 
ll.age                  <- c( 0,   1,   5,   10,  15,  20,  25,  30,  35,  40,  45 , 50,   55,   60,   65,   70,   75,  80,  85)
ul.age                  <- c( 1,   5,   10,  15,  20,  25,  30,  35,  40,  45,  50,  55,   60,   65,   70,   75,   79,  85,  90)
    
i.age_dist              <- c(.04, .05, .05, .06, .06, .06, .06, .07, .07, .07, .06, .06, .06, .06, .06, .04, .04, .02 ,.01) 

#Three step process to add ages by year- 
    #First define upper and lower bounds
   age_ll               <-  c( rep(ll.age, (n*i.age_dist)))
   age_ul               <-  c(rep(ul.age, (n*i.age_dist)))
     
#Then draw n uniform distributions                                     
age_yrs                 <- runif(n, min=age_ll, max=age_ul)


#Network with n people, partnerships are not directed 
net1                    <- network.initialize(n, directed=FALSE)




# # # # # # Add attributed to network  # # # # 
#and a more complex one 
set.vertex.attribute(net1, "sex", v.sex) 
set.vertex.attribute(net1, "behave", v.behave) 
set.vertex.attribute(net1, "position", v.position) 
set.vertex.attribute(net1, "age", round(age_yrs, 1)) 
set.vertex.attribute(net1, "age.actv", v.age_active)
set.vertex.attribute(net1, "sex.actv", ifelse(age_yrs>=v.age_active, 1, 0))


#Check that this is correct
n==(n.msmo + n.msmw + n.mswo + n.wsmo)

```

##Create MSM network 

```{r msm_network}

edges_msm               <- ((n.msmo + n.msmw)* 0.64 * 0.5)       #an average of 0.74 relationships per person -> edges is this times 0.5
nodes_msm_c             <- ((n.msmo + n.msmw)* 0.08)             # 8% of all MSM or MSMW should be in concurrent nodes
nodes_msm_v             <- ((n.msmo + n.msmw)* 0.64 * p.pos_v)   #an average of 0.64 relationships per versatile person 

#For MSM network we will assume straight individuals can have no partnerships- this will create a very sparce network

targets                 <- c(edges_msm,  #total edges
                              (0),        #nodefactor- the number of WSMO nodes that appear in any edge (0) 
                              (0),        #nodefactor- the number of MSWO nodes that appear in any edge (0) 
                              (0),        #node mix- restrict receptive/receptive partnerships (0) 
                              (0),        #node mix- restrict insertive/insertive partnerships (0)  
                              (0),        #nodefactor- those who are not sexually active cannot have partners (0) 
                            nodes_msm_c, #concurrent nodes   
                            nodes_msm_v) #nodefactor- expected number of versatile nodes   
     

## Model fit ---- 
#Dissolution Model
f.coef.diss             <- dissolution_coefs( dissolution = (~offset(edges)),
                                               duration   = c(msm_duration))
                                              #  d.rate    = mean(df.mort_all$Male))

#Formation model
 f.form                 <- ~edges  +  
                           nodefactor("behave", levels=c(3,4))    +
                           nodemix("position", levels2=c(1,6))    + 
                           nodefactor("sex.actv", levels=c(1))    +
                           concurrent                             +
                           nodefactor("position", levels=c(4)) 
	
	
fit_msm                 <- netest(net1,  
                                  formation    = f.form, 
                                  target.stats = targets, 
                                 #edapprox     = FALSE,
                                  coef.diss    = f.coef.diss, 
                              set.control.ergm = control.ergm(MCMC.burnin = 1e4,
                                                             MCMC.interval= 1e4, 
                                                              MCMLE.density.guard = exp(7)), 
                           set.control.stergm = control.network(MCMC.burnin.min = 1e4, MCMC.burnin.max = 1e4))

summary(fit_msm)


```
Run model diagnositics  

```{r model_dx}
start                    <- (Sys.time())
#Simulate our model given the formation and dissolution we've established. Run for 1000 steps but just 1 simulation of it
dx_sim                  <- netdx(fit_msm, nsteps = (1000), 
                                 nsims = 3, 
                                 keep.tnetwork = FALSE, 
                                 ncores = 3, 
                                 keep.tedgelist = TRUE, 
                                 nwstats.formula= ~edges + 
                                                  concurrent + 
                                                  nodefactor("position", levels=c(1,2,3, 4)) + 
                                                  degree(c(0,1,2,3)))

#show what happened as a table
dx_sim$stats.table.formation

plot(dx_sim, type  = "formation" )
plot(dx_sim, type  = "duration")
plot(dx_sim, stats = c("edges", "concurrent"))


#Also test the static version of the model- this help in troubleshooting 

  dx.static     <- netdx(fit_msm, nsims = 1000, dynamic=FALSE)
  dx.static
  
 (end1                    <- (Sys.time() - start))

```

```{r model_dx_burnin1k}

 start                    <- (Sys.time())

#Simulate our model given the formation and dissolution we've established. Run for 1000 steps but just 1 simulation of it
dx_sim1                  <- netdx(fit_msm, nsteps = (1000), 
                                 nsims = 3, 
                                 keep.tnetwork = FALSE, 
                                 ncores = 3, 
                                 keep.tedgelist = TRUE, 
                                 set.control.ergm = control.simulate.ergm(MCMC.burnin = 1e4,
                                                                          MCMC.interval= 1e4), 
                               set.control.stergm = control.simulate.network(MCMC.burnin.min = 1e+4, MCMC.burnin.max = 1e+4),
                                 nwstats.formula= ~edges + 
                                                  concurrent + 
                                                  nodefactor("position", levels=c(1,2,3)) + 
                                                  degree(c(0,1,2,3)))

#show what happened as a table
dx_sim1$stats.table.formation

plot(dx_sim1, type  = "formation" )
plot(dx_sim1, type  = "duration")
plot(dx_sim1, stats = c("edges", "concurrent"))


#Also test the static version of the model- this help in troubleshooting 

#  dx.static     <- netdx(fit_msm, nsims = 1000, dynamic=FALSE)
#  dx.static
  (end2                    <- (Sys.time() - start))
```


```{r model_dx_burnin10k}
 
  start                    <- (Sys.time())
 
 #Simulate our model given the formation and dissolution we've established. Run for 1000 steps but just 1 simulation of it
 dx_sim2                  <- netdx(fit_msm, nsteps = (1000), 
                                  nsims = 3, 
                                  keep.tnetwork = FALSE, 
                                  ncores = 3, 
                                  keep.tedgelist = TRUE, 
                              #    set.control.ergm = control.simulate.ergm(MCMC.burnin = 1e5,
                              #                                             MCMC.interval= 1e5), 
                              #   set.control.stergm = control.simulate.network(MCMC.burnin.min = 0.7e+5, MCMC.burnin.max = 0.7e+5),
                                  nwstats.formula= ~edges + 
                                                   concurrent + 
                                                   nodefactor("position", levels=c(1,2,3)) + 
                                                   degree(c(0,1,2,3)))
 
 #show what happened as a table
 dx_sim2$stats.table.formation
 
 plot(dx_sim2, type  = "formation" )
 plot(dx_sim2, type  = "duration")
 plot(dx_sim2, stats = c("edges", "concurrent"))
 
 
 ##Also test the static version of the model- this help in troubleshooting 
 
 ##  dx.static     <- netdx(fit_msm, nsims = 1000, dynamic=FALSE)
 ##  dx.static
   (end3                    <- (Sys.time() - start))
```
##Create MSW network

```{r msw_network}

edges_msw               <- ((n.mswo + n.wsmo)* 0.75 * 0.5)       #an average of 0.81 relationships per person -> edges is this times 0.5
edges_msw_c             <- ((n.mswo + n.wsmo)* 0.08)              #7% of all MSW should be in concurrent nodes

#For MSM network we will assume straight individuals can have no partnerships- this will create a very sparce network

targets                 <- c(edges_msw,  #total edges
                             (0),        #nodefactor- the number of MSM nodes that appear in any edge (0) 
                             (0),        #node mix- restrict F/F partnerships (0) 
                             (0),        #node mix- restrict M/M partnerships (0)  
                             (0),        #nodefactor- those who are not sexually active cannot have partners (0) 
                            edges_msw_c) #concurrent nodes   

## Model fit ---- 
#Dissolution Model
f.coef.diss             <- dissolution_coefs( dissolution = (~offset(edges)),
                                               duration   = c(msw_duration))
                                              #  d.rate    = mean(df.mort_all$Male))

#Formation model
 f.form                 <- ~edges  +  
                           nodefactor("behave", levels=c(1)) +
                           nodemix("sex", levels2=c(1,3))    + 
                           nodefactor("sex.actv", levels=c(1))  +
                           concurrent 
	
	
fit_msw                 <- netest(net1,  
                                 formation    = f.form, 
                                 target.stats = targets, 
                                 #edapprox     = FALSE,
                                 coef.diss    = f.coef.diss) 

summary(fit_msw)
```

Run model diagnostics 

```{r model_dx_default}
start                   <- Sys.time()
#Simulate our model given the formation and dissolution we've established. Run for 1000 steps but just 1 simulation of it
dx_sim                  <- netdx(fit_msw, nsteps = (1000), 
                                 nsims = 1, 
                                 keep.tnetwork = FALSE, 
                                 ncores = 3, 
                                 keep.tedgelist = TRUE,
                                 nwstats.formula= ~edges + 
                                                  concurrent + 
                                                  degree(c(0,1,2,3)))

#show what happened as a table
dx_sim$stats.table.formation

plot(dx_sim, type  = "formation" )
plot(dx_sim, type  = "duration")
plot(dx_sim, stats = c("edges", "concurrent"))


#Also test the static version of the model- this helps in troubleshooting 

  dx.static     <- netdx(fit_msw, nsims = 1000, dynamic=FALSE)
  dx.static
  
(end                    <- (Sys.time() - start))
```



Tester infection model for MSM

```{r infection_model}

  source("./09_initialize_function_test.R")
  source("./10_resim_function.R")

 l.fits                 <- list(fit_msm, fit_msw)


 # Parameterizing an SIS epidemic
 param                  <- param.net(inf.prob = 0.4,
                                     rec.rate = 0.003,
                                         modes=1)
 
 
  # Initial conditions
 init                   <- init.net(i.num = n*.3)
 
 # Control settings
 control                <- control.net(type = "SI", 
                                      nsims = 1, 
                                     epi.by = "behave",
                                      ncores= 8,
                                     nsteps = 10,  #(50*12), 
                             keep.tedgelist = TRUE, 
                             initialize.FUN = initialize_func,
                       #   set.control.ergm = control.simulate.ergm(MCMC.burnin = 1e5,
                      #                                           MCMC.interval= 1e5), 
                      #          set.control.stergm = control.simulate.network(MCMC.burnin.min = 0.7e+5, 
                      #                                                        MCMC.burnin.max = 0.7e+5),
                              keep.tnetwork = TRUE,  #,
                             resim_nets.FUN = simnet_hpv,
                               module.order = c("resim_nets.FUN","departures.FUN","arrivals.FUN","recovery.FUN","edges_correct.FUN","get_prev.FUN") , 
                              # delete.nodes = TRUE, 
                                  skip.check = TRUE,
                                     depend = TRUE,
                                    verbose = FALSE)
          
 
# Run the network model simulation with netsim
sim_short              <- netsim(l.fits, param, init, control)

  
  ## Main network
  nwparam.m            <- EpiModel::get_nwparam(dat, network = 1)
  

plot(sim_short)
```


storing other code from EpiModel
```{r store}

#' @description Adjusts the edges coefficients in a dynamic network model
#'              to preserve the mean degree.
#'
#' @inheritParams aging_msm
#'
#' @details
#' In HIV/STI modeling, there is typically an assumption that changes in
#' population size do not affect one's number of partners, specified as the
#' mean degree for network models. A person would not have 10 times the number
#' of partners should he move from a city 10 times as large. This module uses
#' the adjustment of Krivitsky et al. to adjust the edges coefficients on the
#' three network models to account for varying population size in order to
#' preserve that mean degree.

#edges_correct_msm <- function(dat, at) {
#
#  old.num <- dat$epi$num[at - 1]
#  new.num <- sum(dat$attr$active == 1, na.rm = TRUE)
#  adjust <- log(old.num) - log(new.num)
#
#  coef.form.m <- get_nwparam(dat, network = 1)$coef.form
#  coef.form.m[1] <- coef.form.m[1] + adjust
#  dat$nwparam[[1]]$coef.form <- coef.form.m
#
#  coef.form.p <- get_nwparam(dat, network = 2)$coef.form
#  coef.form.p[1] <- coef.form.p[1] + adjust
#  dat$nwparam[[2]]$coef.form <- coef.form.p
#
#  coef.form.i <- get_nwparam(dat, network = 3)$coef.form
#  coef.form.i[1] <- coef.form.i[1] + adjust
#  dat$nwparam[[3]]$coef.form <- coef.form.i
#
#  return(dat)
#}


calc_resim_nwstats <- function(dat, at) {

  for (nw in 1:3) {
    n <- attr(dat$el[[nw]], "n")
    edges <- nrow(dat$el[[nw]])
    meandeg <- round(edges / n, 3)
    concurrent <- round(mean(get_degree(dat$el[[nw]]) > 1), 3)
    mat <- matrix(c(edges, meandeg, concurrent), ncol = 3, nrow = 1)
    if (at == 2) {
      dat$stats$nwstats[[nw]] <- mat
      colnames(dat$stats$nwstats[[nw]]) <- c("edges", "meand", "conc")
    }
    if (at > 2) {
      dat$stats$nwstats[[nw]] <- rbind(dat$stats$nwstats[[nw]], mat)
    }
  }

  return(dat)
}

#A function to analyze my own networks based on EpiModel HIV code

#```{r creation_network_function}



## The base code: 

resim_nets <- function(dat, at) {

  browser()
  
  idsActive <- which(dat$attr$active == 1)
  anyActive <- ifelse(length(idsActive) > 0, TRUE, FALSE)
  if (dat$param$modes == 2) {
    nActiveM1 <- length(intersect(modeids(dat$nw, mode = 1), idsActive))
    nActiveM2 <- length(intersect(modeids(dat$nw, mode = 2), idsActive))
    anyActive <- ifelse(nActiveM1 > 0 & nActiveM2 > 0, TRUE, FALSE)
  }

  # Pull network model parameters
  nwparam <- get_nwparam(dat)

  # Serosorting model check
  statOnNw <- ("status" %in% dat$temp$fterms)
  status <- dat$attr$status
  if (statOnNw == TRUE && length(unique(status)) == 1) {
    stop("Stopping simulation because status in formation formula and ",
         "no longer any discordant nodes",
         call. = TRUE)
  }

  # Set up nwstats df
  if (dat$control$save.nwstats == TRUE) {
    if (at == 2) {
      nwstats <- attributes(dat$nw)$stats
      dat$stats$nwstats <- as.data.frame(nwstats)
    }
  }

  # Network simulation
  if (anyActive > 0 & dat$control$depend == TRUE) {
    suppressWarnings(
      dat$nw <- simulate(dat$nw,
                         formation = nwparam$formation,
                         dissolution = nwparam$coef.diss$dissolution,
                         coef.form = nwparam$coef.form,
                         coef.diss = nwparam$coef.diss$coef.adj,
                         constraints = nwparam$constraints,
                         time.start = at,
                         time.slices = 1,
                         time.offset = 0,
                         monitor = dat$control$nwstats.formula,
                         control = dat$control$set.control.stergm))

    # Set up nwstats df
    if (dat$control$save.nwstats == TRUE) {
      dat$stats$nwstats <- rbind(dat$stats$nwstats,
                                 tail(attributes(dat$nw)$stats, 1)[,])
    }

    if (dat$control$delete.nodes == TRUE) {
      dat$nw <- network.extract(dat$nw, at = at)
      inactive <- which(dat$attr$active == 0)
      dat$attr <- deleteAttr(dat$attr, inactive)
    }

  }

  return(dat)
}




```

```{r init_func}


```
